refactor: consolidate webhook routes
