+++ backend/routes/webhooksRoutes.ts
@@ -1,5 +1,7 @@
 import express from 'express';
 import crypto from 'crypto';
+import { handleWorkOrderHook } from '../controllers/WebhookController';
+import { requireThirdPartyAuth } from '../middleware/thirdPartyAuth';
 import Webhook from '../models/Webhook';
 import idempotency from '../middleware/idempotency';
 
@@ -15,5 +17,8 @@ router.post('/subscribe', idempotency, async (req, res) => {
   const hook = await Webhook.create({ url, event, secret });
   res.status(201).json({ id: hook._id, url: hook.url, event: hook.event, secret });
 });
+
+// Work order webhook
+router.post('/workorder', requireThirdPartyAuth, handleWorkOrderHook);
 
 export default router;
